% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{list_parameters}
\alias{list_parameters}
\title{List all available chemical parameters}
\usage{
list_parameters()
}
\value{
A character vector of parameter names.
}
\description{
Get complete list of all available chemical parameters in the database.
}
\examples{
\dontrun{
# Get all parameters
params <- list_parameters()
}
}
