% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-client.R
\name{list_sources}
\alias{list_sources}
\title{List all guideline sources and documents}
\usage{
list_sources()
}
\value{
A list of sources with nested document information.
}
\description{
Get list of all guideline sources and their associated documents.
}
\examples{
\dontrun{
# Get all sources
sources <- list_sources()
}
}
